+++ Assets/coherence/baked/Interop/_01ab9e945591dab4c9c7a936045b976c_2272714898920472812.cs
@@ -45,7 +45,7 @@ namespace Coherence.Generated
 
             var comp = (Interop*)data;
 
-            orig.ids = new byte[comp->ids.Length]; Marshal.Copy((System.IntPtr)comp->ids.Data, orig.ids, 0, comp->ids.Length);
+            orig.ids = new byte[comp->ids.Length]; Marshal.Copy((System.IntPtr)comp->ids.Data, orig.ids, 0, (int)comp->ids.Length);
 
             return orig;
         }
